Scraper/requirements.txt
crawl4ai>=0.4.9
fastapi>=0.115.5,<1.0.0
pydantic>=2.9.0,<3.0.0
python-dotenv>=1.0.1
uvicorn[standard]>=0.32.0,<1.0.0
playwright>=1.49.0
pymongo>=4.10.1